feat(registry): add pi-rust agent, remove pi-lite (restored from archived a3)
Preserved from the archived a3 working copy (branch 04-03-chore_pi_rust):
replaces the pi-lite agent with a new registry/agent/pi-rust agent; updates
benchmark scripts and package.json. Includes the pre-existing 'chore: pi rust'
commit.
