--- registry/agent/pi-lite/package.json
@@ -1,35 +1,0 @@
-{
-	"name": "@rivet-dev/agent-os-pi-lite",
-	"version": "0.1.0",
-	"type": "module",
-	"license": "Apache-2.0",
-	"main": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"files": [
-		"dist",
-		"bin"
-	],
-	"bin": {
-		"pi-lite-acp": "./dist/adapter.js",
-		"pi-lite": "./bin/pi-lite"
-	},
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.js",
-			"default": "./dist/index.js"
-		}
-	},
-	"scripts": {
-		"build": "node ./scripts/build-pi-lite.mjs && tsc",
-		"check-types": "tsc --noEmit"
-	},
-	"dependencies": {
-		"@agentclientprotocol/sdk": "^0.16.1",
-		"@rivet-dev/agent-os-core": "workspace:*"
-	},
-	"devDependencies": {
-		"@types/node": "^22.10.2",
-		"typescript": "^5.7.2"
-	}
-}
