+++ registry/agent/pi-rust/pi-rust
@@ -1,0 +1,1 @@
+../../registry/agent/pi-rust
